pet/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.0
pyyaml>=6.0
httpx>=0.27
anthropic>=0.40
openai>=1.0
fastmcp>=2.0
